[Restructuring] API-AUTH
