+++ src/endpoints/auth/__init__.py
@@ -1,0 +1,5 @@
+from src.endpoints.auth.post.create_new_librarian import create_new_librarian
+from src.endpoints.auth.post.create_new_user import create_new_user
+from src.endpoints.auth.post.login import login_for_access_token
+from src.endpoints.auth.post.logout import logout
+from src.endpoints.auth.router_init import router
