+++ src/endpoints/auth/router_init.py
@@ -1,0 +1,5 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    prefix="/auth", tags=["auth"], responses={401: {"user": "Not authorized"}}
+)
